sharing fixture with multiple modules
