+++ services/director-v2/tests/unit/conftest.py
@@ -3,7 +3,9 @@
 
 import random
 
+import aiodocker
 import pytest
+import tenacity
 from _pytest.monkeypatch import MonkeyPatch
 from dask.distributed import LocalCluster, Scheduler, Worker
 from distributed.deploy.spec import SpecCluster
@@ -148,3 +150,39 @@ def dask_spec_local_cluster(
     monkeypatch.setenv("DASK_SCHEDULER_PORT", f"{scheduler_address.port}")
     yield cluster
     cluster.close()
+
+
+@pytest.fixture
+async def docker_swarm(loop) -> None:
+    # NOTE: overwrites existing fixture
+    class _NotInSwarmException(Exception):
+        pass
+
+    async def _in_docker_swarm(raise_error: bool = False) -> bool:
+        try:
+            inspect_result = await docker.swarm.inspect()
+            assert type(inspect_result) == dict
+        except aiodocker.exceptions.DockerError as error:
+            assert error.status == 503
+            assert error.message.startswith("This node is not a swarm manager")
+            if raise_error:
+                raise _NotInSwarmException() from error
+            return False
+        return True
+
+    async with aiodocker.Docker() as docker:
+        async for attempt in tenacity.AsyncRetrying(
+            wait=tenacity.wait_exponential(), stop=tenacity.stop_after_delay(20)
+        ):
+            with attempt:
+                if not await _in_docker_swarm():
+                    await docker.swarm.init()
+                # if still not in swarm raises an error to try and initialize again
+                await _in_docker_swarm(raise_error=True)
+
+        assert await _in_docker_swarm() is True
+
+        yield
+
+        await docker.swarm.leave(force=True)
+        assert await _in_docker_swarm() is False
